desktop_app/requirements.txt
PySide6>=6.6
requests>=2.31
python-dotenv>=1.0
